@@ -103,9 +103,11 @@ func (cli *PipelineRunner) TearDown() {
 		log.Errorf("failed on tear-down deleting the config-map %s, ns: %s. %v", cli.pipelineInfo.RadixConfigMapName, namespace, err)
 	}
 
-	err = cli.kubeUtil.DeleteConfigMap(namespace, cli.pipelineInfo.GitConfigMapName)
-	if err != nil {
-		log.Errorf("failed on tear-down deleting the config-map %s, ns: %s. %v", cli.pipelineInfo.GitConfigMapName, namespace, err)
+	if cli.pipelineInfo.PipelineArguments.PipelineType == string(v1.BuildDeploy) {
+		err = cli.kubeUtil.DeleteConfigMap(namespace, cli.pipelineInfo.GitConfigMapName)
+		if err != nil {
+			log.Errorf("failed on tear-down deleting the config-map %s, ns: %s. %v", cli.pipelineInfo.GitConfigMapName, namespace, err)
+		}
 	}
 }
 
