@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/equinor/radix-common/utils/errors"
 	errorUtils "github.com/equinor/radix-common/utils/errors"
 	"github.com/equinor/radix-operator/pipeline-runner/model"
 	application "github.com/equinor/radix-operator/pkg/apis/applicationconfig"
@@ -15,7 +14,7 @@ import (
 	"github.com/equinor/radix-operator/pkg/apis/utils"
 	radixclient "github.com/equinor/radix-operator/pkg/client/clientset/versioned"
 	log "github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 	corev1 "k8s.io/api/core/v1"
 )
 
@@ -80,30 +79,31 @@ func (cli *ApplyConfigStepImplementation) Run(pipelineInfo *model.PipelineInfo) 
 	// Set back to pipeline
 	pipelineInfo.SetApplicationConfig(applicationConfig)
 
-	gitConfigMap, err := cli.GetKubeutil().GetConfigMap(namespace, pipelineInfo.GitConfigMapName)
-	if err != nil {
-		log.Errorf("could not retrieve git values from temporary configmap %s, %v", pipelineInfo.GitConfigMapName, err)
-		return nil
+	if pipelineInfo.PipelineArguments.PipelineType == string(v1.BuildDeploy) {
+		gitCommitHash, gitTags := cli.getHashAndTags(namespace, pipelineInfo)
+		pipelineInfo.SetGitAttributes(gitCommitHash, gitTags)
 	}
-
-	//lint:ignore SA4006 temporary fix
-	gitCommitHash, commitErr := getValueFromConfigMap(defaults.RadixGitCommitHashKey, gitConfigMap)
-
-	// TODO: Remove this line once radix-tekton writes the correct commitId to the configmap in the line above
-	gitCommitHash = pipelineInfo.PipelineArguments.CommitID
-
-	gitTags, tagsErr := getValueFromConfigMap(defaults.RadixGitTagsKey, gitConfigMap)
-	err = errorUtils.Concat([]error{commitErr, tagsErr})
-	if err != nil {
-		log.Errorf("could not retrieve git values from temporary configmap %s, %v", pipelineInfo.GitConfigMapName, err)
-		return nil
-	}
-	pipelineInfo.SetGitAttributes(gitCommitHash, gitTags)
 
 	return nil
 }
 
-//CreateRadixApplication Create RadixApplication from radixconfig.yaml content
+func (cli *ApplyConfigStepImplementation) getHashAndTags(namespace string, pipelineInfo *model.PipelineInfo) (string, string) {
+	gitConfigMap, err := cli.GetKubeutil().GetConfigMap(namespace, pipelineInfo.GitConfigMapName)
+	if err != nil {
+		log.Errorf("could not retrieve git values from temporary configmap %s, %v", pipelineInfo.GitConfigMapName, err)
+		return "", ""
+	}
+	gitCommitHash, commitErr := getValueFromConfigMap(defaults.RadixGitCommitHashKey, gitConfigMap)
+	gitTags, tagsErr := getValueFromConfigMap(defaults.RadixGitTagsKey, gitConfigMap)
+	err = errorUtils.Concat([]error{commitErr, tagsErr})
+	if err != nil {
+		log.Errorf("could not retrieve git values from temporary configmap %s, %v", pipelineInfo.GitConfigMapName, err)
+		return "", ""
+	}
+	return gitCommitHash, gitTags
+}
+
+// CreateRadixApplication Create RadixApplication from radixconfig.yaml content
 func CreateRadixApplication(radixClient radixclient.Interface,
 	configFileContent string) (*v1.RadixApplication, error) {
 	ra := &v1.RadixApplication{}
@@ -125,7 +125,7 @@ func CreateRadixApplication(radixClient radixclient.Interface,
 	isRAValid, errs := validate.CanRadixApplicationBeInsertedErrors(radixClient, ra)
 	if !isRAValid {
 		log.Errorf("Radix config not valid.")
-		return nil, errors.Concat(errs)
+		return nil, errorUtils.Concat(errs)
 	}
 	return ra, nil
 }
