@@ -3,6 +3,7 @@ package steps
 import (
 	"context"
 	"fmt"
+	"github.com/equinor/radix-operator/pkg/apis/defaults"
 
 	"github.com/equinor/radix-operator/pipeline-runner/model"
 	"github.com/equinor/radix-operator/pkg/apis/deployment"
@@ -159,20 +160,21 @@ func areArgumentsValid(arguments model.PipelineArguments) error {
 }
 
 func mergeWithRadixApplication(radixConfig *v1.RadixApplication, radixDeployment *v1.RadixDeployment, environment string) error {
-	if err := mergeComponentsWithRadixApplication(radixConfig, radixDeployment, environment); err != nil {
+	defaultEnvVars := getDefaultEnvVarsFromRadixDeployment(radixDeployment)
+	if err := mergeComponentsWithRadixApplication(radixConfig, radixDeployment, environment, defaultEnvVars); err != nil {
 		return err
 	}
 
-	if err := mergeJobComponentsWithRadixApplication(radixConfig, radixDeployment, environment); err != nil {
+	if err := mergeJobComponentsWithRadixApplication(radixConfig, radixDeployment, environment, defaultEnvVars); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-func mergeJobComponentsWithRadixApplication(radixConfig *v1.RadixApplication, radixDeployment *v1.RadixDeployment, environment string) error {
+func mergeJobComponentsWithRadixApplication(radixConfig *v1.RadixApplication, radixDeployment *v1.RadixDeployment, environment string, defaultEnvVars v1.EnvVarsMap) error {
 	newEnvJobs := deployment.
-		NewJobComponentsBuilder(radixConfig, environment, make(map[string]pipeline.ComponentImage), make(v1.EnvVarsMap)).
+		NewJobComponentsBuilder(radixConfig, environment, make(map[string]pipeline.ComponentImage), defaultEnvVars).
 		JobComponents()
 
 	newEnvJobsMap := make(map[string]v1.RadixDeployJobComponent)
@@ -194,8 +196,8 @@ func mergeJobComponentsWithRadixApplication(radixConfig *v1.RadixApplication, ra
 	return nil
 }
 
-func mergeComponentsWithRadixApplication(radixConfig *v1.RadixApplication, radixDeployment *v1.RadixDeployment, environment string) error {
-	newEnvComponents, err := deployment.GetRadixComponentsForEnv(radixConfig, environment, make(map[string]pipeline.ComponentImage), make(v1.EnvVarsMap))
+func mergeComponentsWithRadixApplication(radixConfig *v1.RadixApplication, radixDeployment *v1.RadixDeployment, environment string, defaultEnvVars v1.EnvVarsMap) error {
+	newEnvComponents, err := deployment.GetRadixComponentsForEnv(radixConfig, environment, make(map[string]pipeline.ComponentImage), defaultEnvVars)
 	if err != nil {
 		return nil
 	}
@@ -217,4 +219,15 @@ func mergeComponentsWithRadixApplication(radixConfig *v1.RadixApplication, radix
 	}
 
 	return nil
+}
+
+func getDefaultEnvVarsFromRadixDeployment(radixDeployment *v1.RadixDeployment) v1.EnvVarsMap {
+	envVarsMap := make(v1.EnvVarsMap)
+	if gitCommitHash, ok := radixDeployment.Annotations[kube.RadixCommitAnnotation]; ok {
+		envVarsMap[defaults.RadixCommitHashEnvironmentVariable] = gitCommitHash
+	}
+	if gitTags, ok := radixDeployment.Annotations[kube.RadixGitTagsAnnotation]; ok {
+		envVarsMap[defaults.RadixGitTagsEnvironmentVariable] = gitTags
+	}
+	return envVarsMap
 }
